// ui/src/lib/blank-modal/interface/modal-interface.ts
export interface IModalItem {
    title: string;
    subTitle: string;
    description: string;
    buttonLeft: string;
    buttonRight: string;
    type: string;
    data?: any;
}
export interface IModalItemDoc {
    title: string;
    subTitle: string;
    description: string;
    buttonLeft: string;
    buttonRight: string;
    type: string;
    tipoData:string;
}